Link both json and yaml
in theory
